bugfix: mplayer not actually restarting
When the darkice encoding stops the mplayer processes stop. The code is supposed to kill all remaining mplayer processes and restart, however, an infinite loop is happening causing the application to stop responding properly. So instead of uising the self method for is_playing in the stop method, statically assign false to self.is_running during the stop() method.